Revert "Downsized packet sizes"
This reverts commit 8611c0e504be70dae5cc59c8dd5cbcd1d4e9c417.

+++ Multiplayer_CA2/Multiplayer_CA2/MultiplayerGameState.cpp
@@ -123,7 +123,7 @@ void MultiplayerGameState::onDestroy()
 	{
 		// Inform server this client is dying
 		sf::Packet packet;
-		packet << static_cast<sf::Int8>(Client::Quit);
+		packet << static_cast<sf::Int32>(Client::Quit);
 		mSocket.send(packet);
 	}
 }
@@ -182,7 +182,7 @@ bool MultiplayerGameState::update(sf::Time dt)
 		if (mSocket.receive(packet) == sf::Socket::Done)
 		{
 			mTimeSinceLastPacket = sf::seconds(0.f);
-			sf::Int8 packetType;
+			sf::Int32 packetType;
 			packet >> packetType;
 			handlePacket(packetType, packet);
 		}
@@ -212,8 +212,8 @@ bool MultiplayerGameState::update(sf::Time dt)
 		while (mWorld.pollGameAction(gameAction))
 		{
 			sf::Packet packet;
-			packet << static_cast<sf::Int8>(Client::GameEvent);
-			packet << static_cast<sf::Int8>(gameAction.type);
+			packet << static_cast<sf::Int32>(Client::GameEvent);
+			packet << static_cast<sf::Int32>(gameAction.type);
 			packet << gameAction.position.x;
 			packet << gameAction.position.y;
 
@@ -224,14 +224,14 @@ bool MultiplayerGameState::update(sf::Time dt)
 		if (mTickClock.getElapsedTime() > sf::seconds(1.f / 20.f))
 		{
 			sf::Packet positionUpdatePacket;
-			positionUpdatePacket << static_cast<sf::Int8>(Client::PositionUpdate);
-			positionUpdatePacket << static_cast<sf::Int8>(mLocalPlayerIdentifiers.size());
+			positionUpdatePacket << static_cast<sf::Int32>(Client::PositionUpdate);
+			positionUpdatePacket << static_cast<sf::Int32>(mLocalPlayerIdentifiers.size());
 
-			FOREACH(sf::Int8 identifier, mLocalPlayerIdentifiers)
+			FOREACH(sf::Int32 identifier, mLocalPlayerIdentifiers)
 			{
 				//Now also passes rotation to keep the tanks synced across the network
 				if (Tank* tank = mWorld.getTank(identifier))
-					positionUpdatePacket << identifier << tank->getPosition().x << tank->getPosition().y << static_cast<sf::Int8>(tank->getHitpoints()) << tank->getRotation();
+					positionUpdatePacket << identifier << tank->getPosition().x << tank->getPosition().y << static_cast<sf::Int32>(tank->getHitpoints()) << tank->getRotation();
 			}
 
 			mSocket.send(positionUpdatePacket);
@@ -255,7 +255,7 @@ void MultiplayerGameState::disableAllRealtimeActions()
 {
 	mActiveState = false;
 
-	FOREACH(sf::Int8 identifier, mLocalPlayerIdentifiers)
+	FOREACH(sf::Int32 identifier, mLocalPlayerIdentifiers)
 		mPlayers[identifier]->disableAllRealtimeActions();
 }
 
@@ -343,7 +343,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// Sent by the server to order to spawn player 1 airplane on connect
 	case Server::SpawnSelf:
 	{
-		sf::Int8 TankIdentifier;
+		sf::Int32 TankIdentifier;
 		sf::Vector2f TankPosition;
 		packet >> TankIdentifier >> TankPosition.x >> TankPosition.y;
 
@@ -359,7 +359,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// 
 	case Server::PlayerConnect:
 	{
-		sf::Int8 TankIdentifier;
+		sf::Int32 TankIdentifier;
 		sf::Vector2f TankPosition;
 		packet >> TankIdentifier >> TankPosition.x >> TankPosition.y;
 
@@ -372,7 +372,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// 
 	case Server::PlayerDisconnect:
 	{
-		sf::Int8 TankIdentifier;
+		sf::Int32 TankIdentifier;
 		packet >> TankIdentifier;
 
 		mWorld.removeTank(TankIdentifier);
@@ -382,7 +382,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// 
 	case Server::InitialState:
 	{
-		sf::Int8 TankCount;
+		sf::Int32 TankCount;
 		float worldHeight, currentScroll;
 		packet >> worldHeight >> currentScroll;
 
@@ -390,11 +390,11 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 		mWorld.setCurrentBattleFieldPosition(currentScroll);
 
 		packet >> TankCount;
-		for (sf::Int8 i = 0; i < TankCount; ++i)
+		for (sf::Int32 i = 0; i < TankCount; ++i)
 		{
-			sf::Int8 TankIdentifier;
-			sf::Int8 hitpoints;
-			sf::Int8 missileAmmo;
+			sf::Int32 TankIdentifier;
+			sf::Int32 hitpoints;
+			sf::Int32 missileAmmo;
 			sf::Vector2f TankPosition;
 			packet >> TankIdentifier >> TankPosition.x >> TankPosition.y >> hitpoints >> missileAmmo;
 
@@ -410,7 +410,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	//
 	case Server::AcceptCoopPartner:
 	{
-		sf::Int8 TankIdentifier;
+		sf::Int32 TankIdentifier;
 		packet >> TankIdentifier;
 
 		mWorld.addTank(TankIdentifier);
@@ -421,7 +421,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// Player event (like missile fired) occurs
 	case Server::PlayerEvent:
 	{
-		sf::Int8 TankIdentifier;
+		sf::Int32 TankIdentifier;
 		sf::Int32 action;
 		packet >> TankIdentifier >> action;
 
@@ -433,7 +433,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// Player's movement or fire keyboard state changes
 	case Server::PlayerRealtimeChange:
 	{
-		sf::Int8 TankIdentifier;
+		sf::Int32 TankIdentifier;
 		sf::Int32 action;
 		bool actionEnabled;
 		packet >> TankIdentifier >> action >> actionEnabled;
@@ -452,7 +452,7 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	// Pickup created
 	case Server::SpawnPickup:
 	{
-		sf::Int8 type;
+		sf::Int32 type;
 		sf::Vector2f position;
 		packet >> type >> position.x >> position.y;
 
@@ -463,15 +463,15 @@ void MultiplayerGameState::handlePacket(sf::Int32 packetType, sf::Packet& packet
 	case Server::UpdateClientState:
 	{
 		float currentWorldPosition;
-		sf::Int8 TankCount;
+		sf::Int32 TankCount;
 		packet >> currentWorldPosition >> TankCount;
 
 		float currentViewPosition = mWorld.getViewBounds().top + mWorld.getViewBounds().height;
 
-		for (sf::Int8 i = 0; i < TankCount; ++i)
+		for (sf::Int32 i = 0; i < TankCount; ++i)
 		{
 			sf::Vector2f TankPosition;
-			sf::Int8 TankIdentifier;
+			sf::Int32 TankIdentifier;
 			float TankRotation;
 			packet >> TankIdentifier >> TankPosition.x >> TankPosition.y >> TankRotation;
